changed it so claims component uses more ssr
