+++ next-client/src/components/claim/index.ts
@@ -1,0 +1,2 @@
+export * from "./Claim";
+export * from "./QuoteCard";
